+++ server/src/main/java/io/littlehorse/common/model/repartitioncommand/repartitionsubcommand/UpdateCountedTagModel.java
@@ -3,8 +3,8 @@ package io.littlehorse.common.model.repartitioncommand.repartitionsubcommand;
 import com.google.protobuf.Empty;
 import com.google.protobuf.Message;
 import io.littlehorse.common.LHServerConfig;
+import io.littlehorse.common.model.CountedTagModel;
 import io.littlehorse.common.model.corecommand.CoreSubCommand;
-import io.littlehorse.common.model.corecommand.subcommand.CountedTagModel;
 import io.littlehorse.common.proto.UpdateCountedTag;
 import io.littlehorse.sdk.common.exception.LHSerdeException;
 import io.littlehorse.server.streams.stores.TenantScopedStore;
