+++ server/src/main/java/io/littlehorse/server/streams/BackendInternalComms.java
@@ -21,7 +21,7 @@ import io.littlehorse.common.LHServerConfig;
 import io.littlehorse.common.Storeable;
 import io.littlehorse.common.exceptions.LHApiException;
 import io.littlehorse.common.model.AbstractGetable;
-import io.littlehorse.common.model.corecommand.subcommand.CountedTagModel;
+import io.littlehorse.common.model.CountedTagModel;
 import io.littlehorse.common.model.getable.ObjectIdModel;
 import io.littlehorse.common.model.getable.core.events.WorkflowEventModel;
 import io.littlehorse.common.model.getable.core.taskworkergroup.HostModel;
